Avoid modifying rpmbuild system variables
The previous SPEC file modified the system variables _bindir and its
friends to include the generated files.  This caused rpmbuild to run
with an unexpected environment and stop with an unclear error.  Do not
modify _bindir, _libdir, or _datadir; instead, use pg_bindir,
pg_libdir, and pg_datadir to avoid interfering with rpmbuild.

(internal: 75a4946)